django_backend/requirements.txt
asgiref==3.9.1
dj-database-url==2.1.0
Django==5.2.6
django-cors-headers==4.7.0
djangorestframework==3.16.1
gunicorn==22.0.0
pillow==11.3.0
psycopg2-binary==2.9.9
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.6.0